Various bug fixes and add missing functions
