+++ src/parsers/ConferenceParser.ts
@@ -1,4 +1,5 @@
 import config, { AvailableBackends } from "../config";
+import { RoomKind } from "../models/room_kinds";
 import { IAuditorium, IConference, IInterestRoom, IPerson, ITalk } from "../models/schedule";
 import { PentabarfParser } from "./PentabarfParser";
 import PretalxParser from "./PretalxParser";
@@ -25,3 +26,19 @@ export const getConferenceParser = async (): Promise<ConferenceParser> => {
         }
     }
 };
+
+export function deprefix(id: string): { kind: RoomKind, name: string; } {
+    const override = config.conference.prefixes.nameOverrides[id];
+
+    const auditoriumPrefix = config.conference.prefixes.auditoriumRooms.find(p => id.startsWith(p));
+    if (auditoriumPrefix) {
+        return { kind: RoomKind.Auditorium, name: override || id.slice(auditoriumPrefix.length) };
+    }
+
+    const interestPrefix = config.conference.prefixes.interestRooms.find(p => id.startsWith(p));
+    if (interestPrefix) {
+        return { kind: RoomKind.SpecialInterest, name: override || id.slice(interestPrefix.length) };
+    }
+
+    return { kind: RoomKind.SpecialInterest, name: override || id };
+}
